Bug fix and typo fixed
Fixed a typo in the web display.
Chnaged varible type from uint8_t to uint16_t to allow bigger values for pump frequency

+++ Terrarium_Control/Terrarium_Control.ino
@@ -52,7 +52,7 @@ struct relayState_t {
 } relayState;
 
 struct settings_t {
-  uint8_t pumpFrequency;          //Pump frequency in minutes
+  uint16_t pumpFrequency;          //Pump frequency in minutes
   uint8_t pumpDuration;           //Number of seconds to run the pump
   uint8_t fanDuration;            //Number of minutes to run the fan
   uint8_t fanFrequency;           //Fan frequency in minutes
